Rework phario repository
